Alias Lookup#sameKey as =#=
This makes it easier to change the order of operands, which may be necessary for it to typecheck in some cases.

For instance, given the following code:

xs.exists(_.lookup.sameKey(myLookup))

If x.lookup is a Lookup[Int, A] and myLookup is a Lookup[Int, B], then if A <: B this will compile.

But if B <: A it will need to be changed to one of the following

xs.exists(x => myLookup.sameKey(x.entity))  // can't use _ shorthand
xs.exists(myLookup sameKey _.entity)        // Scala 3 doesn't like infix alphabetic methods, unidiomatic
xs.exists(myLookup =#= _.entity)            // now possible

Additionally, editors (such as IntelliJ) may have a hotkey to change the order of operands, which may not work for
alphabetic methods, even in infix position.